Worked with 2 kind of forms.
One for refill some data about a person, and a click button who pass the data to the second one.

And the second one form (as a showDialog) has 2 buttons, an Okay button and Cancel button.
Depends of what button you click, it shows a message in the output with the state of the Dialog.